Add files via upload
This is the latest updates to the home page of the website plus the services page. I am currently trying to sort out the dark mode feature so using version control to go back to older versions to see how the code has developed and see if its better or worse then originally.